+++ packages/tools/create-dlightjs/templates/dlight-vite-min-js/vite.config.js
@@ -3,6 +3,6 @@ import dlight from "vite-plugin-dlight"
 
 export default defineConfig({
     plugins: [
-        dlight({ files: "**/*.{view,model}.ts" })
+        dlight({ files: "**/*.{view,model}.js" })
     ]
 });